refactor: project_detail_page 위젯 분리
